Fix reordering actions sometimes causing crashes
The problem was that drag-and-drop was attached to the underlying GTK
widget multiple times (which occurs when e.g. switching between
procedures in ActionBrowser), and so the drag-and-drop events were
triggered for `ActionItem`s that were no longer present in the list of
procedures/constraints (as a result of the switching, where a new
temporary `ActionItem` is added and the previous one is removed).