setting: Remove empty values
The concept of "empty setting values" was made obsolete with the recent
changes to the GIMP 3.0 API (e.g. allowing `None` as a valid value for
GIMP images, layers, resources, ...), optional preprocessing before a
value is assigned to a setting (sometimes making the value valid if it
is not) and generally the lack of usage in the code (including the
`ChoiceSetting` empty values).