Fix export for image-wide procedures and refactor
A new image copy is now created for each layer. Multi-layer images are
also handled accordingly. The performance hit for duplicating an empty
image for each layer is negligible.

The optimization in `Insert back/foreground` involving keeping a merged
copy of the back/foreground layer is no longer applicable as the layer
copy is attached to an image copy that no longer exists for subsequent
layers.