Finish reworking handling of `ItemTree` items in settings
With this change, we now finally have a unified interface for files,
GIMP images and GIMP items in the GUI. This allows to reuse preview
widgets for Convert to a large extent and generally reduces the number
of code changes required in other places.

This also paves way for exporting/editing layers in multiple opened
images, not just the currently displayed image, if that need arises.