DimensionBox: Keep separate values for pixels, percentages and other units
This is more consistent with GIMP which automatically recalculates
values in the Scale tool based on inches and image resolution (values of
 percentages are stored separately and are not recalculated). Since the
actual image resolution in Batcher is only known during the processing,
pixels cannot be recalculated based on inches (a prior procedure or
custom plug-in could have changed the resolution), and so we keep a
separate value for pixels and inches (and percentages as well).

This is also much safer since e.g. changing 5 inches to 5 meters (a much
 bigger dimension) could clog the RAM and crash the plug-in.