DragAndDropContext: Suppress existing drag-and-drop functionality
For `ArrayBox`, this caused e.g. dragging `Gtk.SpinButton`s (in order to
 reorder them) to trigger drag-and-drop twice, causing reordering to
also be activated twice. The reordered widget then ended up in the wrong
 position.

This effectively disables any existing drag-and-drop functionality for
each `ItemBox` child widget. If it turns out there are child widgets
whose DND functionality is indispensable, this change will be revised.